cv/requirements.txt
weasyprint
markdown
pyyaml
jinja2
